--- src/renderer/src/components/terminal-pane/terminal-visibility-effect-phase.ts
@@ -1,7 +1,0 @@
-export type TerminalVisibilityEffectPhase = 'layout' | 'passive'
-
-export function getTerminalVisibilityEffectPhase(
-  platform: NodeJS.Platform
-): TerminalVisibilityEffectPhase {
-  return platform === 'darwin' ? 'layout' : 'passive'
-}
