+++ src/renderer/src/components/terminal-pane/terminal-visibility-resume.ts
@@ -30,10 +30,6 @@ type ResumeTerminalVisibilityArgs = {
   withSuppressedScrollTracking: (callback: () => void) => void
 }
 
-export type TerminalVisibilityPostPaintRecovery = {
-  run: (manager: PaneManager) => void
-}
-
 type HideTerminalVisibilityArgs = {
   manager: PaneManager
   wasVisible: boolean
@@ -61,7 +57,7 @@ export function resumeTerminalVisibility({
   shouldUseLightTabResume,
   captureViewportPositions,
   withSuppressedScrollTracking
-}: ResumeTerminalVisibilityArgs): TerminalVisibilityPostPaintRecovery | null {
+}: ResumeTerminalVisibilityArgs): void {
   // Why: hiding the surface fired mouseleave, which cleared xterm's current
   // link but left its hover cell cache; without this reset a link stays dead
   // until a scroll when the pointer returns to the same cell on reveal.
@@ -89,28 +85,19 @@ export function resumeTerminalVisibility({
         focusActivePane(manager)
       }
     } else {
-      resumeTerminalVisibilityBeforePaint(manager, isActive)
+      resumeTerminalVisibilityHeavy(manager, isActive)
     }
     enforceTerminalViewportIntents(manager)
-    if (shouldUseLightTabResume) {
-      manager.scheduleRevealRepaint()
+    if (!shouldUseLightTabResume) {
+      // Why: this clear wipes the glyph atlas shared with other same-config
+      // terminals; refresh after reset so rebuilt atlases repaint from xterm.
+      resetAndRefreshAllTerminalWebglAtlases()
     }
+    // Why: the synchronous recovery above can fire before the revealed pane is
+    // attached and laid out, where the WebGL renderer drops redraw requests
+    // without retry. Follow up with a settled-frame, pane-scoped repaint.
+    manager.scheduleRevealRepaint()
   })
-  if (shouldUseLightTabResume) {
-    return null
-  }
-  return {
-    run: (currentManager) => {
-      withSuppressedScrollTracking(() => {
-        drainVisibleTerminalBacklog(currentManager)
-        enforceTerminalViewportIntents(currentManager)
-        // Why: the shared atlas reset repaints every live manager, so keep it
-        // out of the reveal's pre-paint critical path.
-        resetAndRefreshAllTerminalWebglAtlases()
-        currentManager.scheduleRevealRepaint()
-      })
-    }
-  }
 }
 
 export function hideTerminalVisibility({
@@ -128,6 +115,8 @@ export function hideTerminalVisibility({
     captureViewportPositions(false)
   }
   if (!isWorktreeActive && (wasVisible || surfaceBecameHidden)) {
+    // Suspend WebGL when going hidden. xterm.write() continues to land in the
+    // DOM-renderer fallback terminal; the suspend is purely a GPU resource decision.
     manager.suspendRendering()
     return { hiddenReason: 'surface', renderingSuspended: true }
   }
@@ -154,13 +143,6 @@ export function recoverVisibleTerminalWindowWake({
 }: RecoverVisibleTerminalWindowWakeArgs): void {
   // Why: macOS screensaver/display wake can leave xterm visible but with a
   // stale renderer/input surface; Orca's own hidden-state resume never runs.
-  // Order: latch intent → resume WebGL → fit metrics → drain → enforce.
-  // Intent must be captured before resume/fit (they can move viewportY and
-  // would re-latch a pin as followOutput). Fit must precede flush so backlog
-  // does not land on the transient DOM↔WebGL one-column-off grid.
-  syncTerminalViewportIntents(manager)
-  manager.resumeRendering()
-  manager.fitAllRevealedPanes()
   for (const pane of manager.getPanes()) {
     requestTerminalBacklogRecovery(pane.terminal)
     flushTerminalOutput(pane.terminal, { maxChars: WINDOW_WAKE_FLUSH_CHARS })
@@ -172,10 +154,10 @@ export function recoverVisibleTerminalWindowWake({
       resetTerminalLinkifierHoverState(pane.terminal)
     }
   }
-  // Why no post-flush re-sync: flushTerminalOutput only submits terminal.write and
-  // returns before parse callbacks, so a same-tick re-sync would read pre-parse
-  // geometry (possibly disturbed by resume/fit) and overwrite the pre-resume pin.
-  // Enforce the pre-resume latched intent instead.
+  syncTerminalViewportIntents(manager)
+  manager.resumeRendering()
+  // Why: wake re-attaches WebGL — same transient cell-metric wobble guard as the heavy resume.
+  manager.fitAllRevealedPanes()
   if (isActive) {
     focusActivePane(manager)
   }
@@ -201,18 +183,27 @@ function requestLightTabBacklogRecovery(manager: PaneManager): void {
   }
 }
 
-function resumeTerminalVisibilityBeforePaint(manager: PaneManager, isActive: boolean): void {
-  manager.resumeRendering()
-  manager.fitAllRevealedPanes()
-  if (isActive) {
-    focusActivePane(manager)
-  }
-}
-
-function drainVisibleTerminalBacklog(manager: PaneManager): void {
+function resumeTerminalVisibilityHeavy(manager: PaneManager, isActive: boolean): void {
+  // Why: hidden panes can accumulate large PTY bursts while Chromium is
+  // occluded. Drain a bounded slice before fitting; the scheduler keeps
+  // ordering and continues the rest asynchronously so return-to-app does
+  // not beachball behind an entire backlog.
   for (const pane of manager.getPanes()) {
     requestTerminalBacklogRecovery(pane.terminal)
     flushTerminalOutput(pane.terminal, { maxChars: VISIBLE_RESUME_FLUSH_CHARS })
+  }
+  syncTerminalViewportIntents(manager)
+  // Resume WebGL immediately so the terminal shows its last-known state
+  // on the first painted frame. macOS context creation is ~5 ms; on
+  // Windows (ANGLE -> D3D11) it can be 100-500 ms but a deferred resume
+  // would paint a stretched DOM-fallback flash, which is worse UX.
+  manager.resumeRendering()
+  // Why: resumeRendering just re-attached WebGL, whose cell metrics briefly differ
+  // from the DOM renderer's; a raw fit here reflows on a transient one-column-off
+  // grid and garbles diff-painting inline TUIs (grok minimize→restore).
+  manager.fitAllRevealedPanes()
+  if (isActive) {
+    focusActivePane(manager)
   }
 }
