+++ src/renderer/src/components/terminal-pane/use-terminal-pane-global-effects.test.ts
@@ -30,8 +30,7 @@ const mocks = vi.hoisted(() => ({
 
 const reactRefState = vi.hoisted(() => ({
   slots: [] as { current: unknown }[],
-  index: 0,
-  effectPhase: null as 'layout' | 'passive' | null
+  index: 0
 }))
 
 function beginHookRender(): void {
@@ -49,21 +48,7 @@ vi.mock('react', async (importOriginal) => {
     ...actual,
     useCallback: <T extends (...args: never[]) => unknown>(callback: T) => callback,
     useEffect: (effect: () => void | (() => void)) => {
-      reactRefState.effectPhase = 'passive'
-      try {
-        effect()
-      } finally {
-        reactRefState.effectPhase = null
-      }
-    },
-    // macOS visibility suspend/resume runs here before reveal paint.
-    useLayoutEffect: (effect: () => void | (() => void)) => {
-      reactRefState.effectPhase = 'layout'
-      try {
-        effect()
-      } finally {
-        reactRefState.effectPhase = null
-      }
+      effect()
     },
     useRef: <T>(value: T) => {
       const index = reactRefState.index
@@ -156,8 +141,7 @@ function useMountForFileDrop(
     isWorktreeActive?: boolean
     isSyncFitEnabled?: boolean
     paneCount?: number
-  } = {},
-  useGlobalEffects: typeof useTerminalPaneGlobalEffects = useTerminalPaneGlobalEffects
+  } = {}
 ): {
   onFileDrop: DropCallback
   manager: {
@@ -171,7 +155,6 @@ function useMountForFileDrop(
     fitAllRevealedPanes: ReturnType<typeof vi.fn>
   }
   paneTransports: Map<number, never>
-  renderingEffectPhases: ('layout' | 'passive' | null)[]
 } {
   let onFileDrop: DropCallback = () => {
     throw new Error('onFileDrop callback was not registered')
@@ -180,10 +163,9 @@ function useMountForFileDrop(
     onFileDrop = callback
     return vi.fn()
   })
-  const renderingEffectPhases: ('layout' | 'passive' | null)[] = []
   const manager = {
     getPanes: vi.fn(() => []),
-    resumeRendering: vi.fn(() => renderingEffectPhases.push(reactRefState.effectPhase)),
+    resumeRendering: vi.fn(),
     resetWebglTextureAtlases: vi.fn(),
     scheduleRevealRepaint: vi.fn(),
     scheduleRevealPresent: vi.fn(),
@@ -194,7 +176,7 @@ function useMountForFileDrop(
   const paneTransports = new Map<number, never>()
 
   beginHookRender()
-  useGlobalEffects({
+  useTerminalPaneGlobalEffects({
     tabId: options.tabId ?? 'tab-1',
     worktreeId: options.worktreeId ?? 'wt-1',
     cwd: options.cwd,
@@ -211,7 +193,7 @@ function useMountForFileDrop(
     toggleExpandPane: vi.fn()
   })
 
-  return { onFileDrop, manager, paneTransports, renderingEffectPhases }
+  return { onFileDrop, manager, paneTransports }
 }
 
 describe('useTerminalPaneGlobalEffects', () => {
@@ -244,26 +226,6 @@ describe('useTerminalPaneGlobalEffects', () => {
     ;(globalThis as unknown as { ResizeObserver: unknown }).ResizeObserver = MockResizeObserver
   })
 
-  it.each([
-    ['darwin', 'layout'],
-    ['win32', 'passive'],
-    ['linux', 'passive']
-  ] as const)('runs visibility transitions in the %s effect phase', async (platform, phase) => {
-    window.api.platform = {
-      get: () => ({ platform, osRelease: 'test', displayServer: null })
-    }
-    vi.resetModules()
-    const { useTerminalPaneGlobalEffects: usePlatformTerminalPaneGlobalEffects } =
-      await import('./use-terminal-pane-global-effects')
-
-    const { renderingEffectPhases } = useMountForFileDrop(
-      { isActive: true, isVisible: true },
-      usePlatformTerminalPaneGlobalEffects
-    )
-
-    expect(renderingEffectPhases).toEqual([phase])
-  })
-
   afterEach(() => {
     for (const manager of registeredManagers.splice(0)) {
       unregisterLivePaneManager(manager)
@@ -274,9 +236,7 @@ describe('useTerminalPaneGlobalEffects', () => {
     delete (globalThis as unknown as { ResizeObserver?: unknown }).ResizeObserver
   })
 
-  it('resumes WebGL and fits before flushing backlog so paint is GPU and grid is stable', () => {
-    // On macOS resume before flush avoids DOM bold flash; fit before flush avoids
-    // writing backlog onto the transient DOM↔WebGL one-column-off grid.
+  it('flushes visible terminal panes before resuming rendering and fitting', () => {
     const order: string[] = []
     const terminalA = { name: 'terminal-a' }
     const terminalB = { name: 'terminal-b' }
@@ -339,23 +299,18 @@ describe('useTerminalPaneGlobalEffects', () => {
     expect(order).toEqual([
       'capture:terminal-a',
       'capture:terminal-b',
-      'resume',
-      'fit-reveal',
-      'intent:terminal-a',
-      'intent:terminal-b',
       'recover:terminal-a',
       'flush:terminal-a',
       'recover:terminal-b',
       'flush:terminal-b',
+      'resume',
+      'fit-reveal',
       'intent:terminal-a',
       'intent:terminal-b',
       'reset-atlas',
       'refresh',
       'reveal-repaint'
     ])
-    // Why: flush must not land between resume and the corrective reveal fit.
-    expect(order.indexOf('resume')).toBeLessThan(order.indexOf('fit-reveal'))
-    expect(order.indexOf('fit-reveal')).toBeLessThan(order.indexOf('flush:terminal-a'))
     expect(mocks.restoreScrollStateAfterLayout).not.toHaveBeenCalled()
     expect(mocks.flushTerminalOutput).toHaveBeenNthCalledWith(1, terminalA, {
       maxChars: 256 * 1024
@@ -366,62 +321,6 @@ describe('useTerminalPaneGlobalEffects', () => {
     expect(mocks.fitPanes).not.toHaveBeenCalled()
     expect(isActiveRef.current).toBe(true)
     expect(isVisibleRef.current).toBe(true)
-  })
-
-  it('records mount-visible completion before PaneManager exists so first tab hide stays light', () => {
-    // Why: PaneManager is created in a passive lifecycle effect after this layout
-    // pass. Bookkeeping must still mark hasCompletedVisibleResume so the first
-    // intra-worktree hide does not take the !hasCompleted suspend branch.
-    const terminal = { name: 'terminal-a' }
-    const manager = {
-      getPanes: vi.fn(() => [{ id: 1, terminal }]),
-      resumeRendering: vi.fn(),
-      resetWebglTextureAtlases: vi.fn(),
-      scheduleRevealRepaint: vi.fn(),
-      scheduleRevealPresent: vi.fn(),
-      refreshAllPanes: vi.fn(),
-      suspendRendering: vi.fn(),
-      fitAllPanes: vi.fn(),
-      fitAllRevealedPanes: vi.fn(),
-      getActivePane: vi.fn(() => null),
-      setActivePane: vi.fn()
-    }
-    registerManagerForReset(manager)
-    const managerRef: { current: typeof manager | null } = { current: null }
-    const baseArgs = {
-      tabId: 'tab-1',
-      worktreeId: 'wt-1',
-      managerRef: managerRef as never,
-      containerRef: { current: null },
-      paneTransportsRef: { current: new Map() },
-      isActiveRef: { current: false },
-      isVisibleRef: { current: false },
-      paneCount: 0,
-      isSyncFitEnabled: true,
-      isWorktreeActive: true,
-      toggleExpandPane: vi.fn()
-    }
-
-    // Mount visible before the manager exists (layout before passive create).
-    beginHookRender()
-    useTerminalPaneGlobalEffects({
-      ...baseArgs,
-      isActive: true,
-      isVisible: true
-    })
-    expect(manager.resumeRendering).not.toHaveBeenCalled()
-
-    // Manager appears; visibility unchanged so the layout effect does not re-run.
-    // First hide still must keep WebGL (light path).
-    managerRef.current = manager
-    beginHookRender()
-    useTerminalPaneGlobalEffects({
-      ...baseArgs,
-      paneCount: 1,
-      isActive: false,
-      isVisible: false
-    })
-    expect(manager.suspendRendering).not.toHaveBeenCalled()
   })
 
   it('uses a light resume for tab switches while the worktree stays active', () => {
